+++ docs/CHANGELOG.md
@@ -1,10 +1,16 @@
 ## Changelogs
+Date: *4/18/2020*
+- Enh: Custom Color
+
+Date: *4/13/2020*
+- Enh: Position Switch Option
+- Chg: New version
+
 Date: *4/9/2020*
 - Enh: Remove JS file completely
 - Fix: Pjax loading issues
 - Enh: Upgraded Version v1.0.1
 - Enh: Updated CHANGELOG.md
-- Enh: Add scroll up to Spaces
 
 Date: *4/4/2020*
 - Enh: Created INSTALLATION.md